requirements.txt
numpy
matplotlib
pyro-ppl
torch
opencv-python
tqdm
mujoco_py
